generic inference job queries
