+++ Dockerfile
@@ -9,7 +9,7 @@
 FROM ubuntu:jammy as rust-base
 
 # NB: This can be "stable" or another version.
-ARG RUST_TOOLCHAIN="1.65.0"
+ARG RUST_TOOLCHAIN="1.66.0"
 
 WORKDIR /tmp
