+++ crates/lib/database_queries/src/queries/generic_inference/job/mod.rs
@@ -1,6 +1,6 @@
 pub mod _keys;
-pub mod list_available_generic_inference_jobs;
 pub mod mark_generic_inference_job_completely_failed;
-pub mod mark_generic_inference_job_successfully_done;
 pub mod mark_generic_inference_job_failure;
 pub mod mark_generic_inference_job_pending_and_grab_lock;
+pub mod mark_generic_inference_job_successfully_done;
+pub mod list_available_generic_inference_jobs;
